+++ src/components/common/PageLayout/index.tsx
@@ -9,7 +9,7 @@ import SideDrawer from './SideDrawer'
 import { useIsSidebarRoute } from '@/hooks/useIsSidebarRoute'
 import { TxModalContext } from '@/components/tx-flow'
 import BatchSidebar from '@/components/batch/BatchSidebar'
-import SocialLoginDeprecation from '@/components/dashboard/SocialLoginDeprecation'
+import SocialLoginDeprecation from '@/components/common/SocialLoginDeprecation'
 
 const PageLayout = ({ pathname, children }: { pathname: string; children: ReactElement }): ReactElement => {
   const [isSidebarRoute, isAnimated] = useIsSidebarRoute(pathname)
